refactor: move logic out of Files into useDisplayOptions hook
